requirements.txt
asgiref==3.6.0
cryptocode==0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pycryptodomex==3.17
pytz==2022.7.1
sqlparse==0.4.3
